feat: implement task controller routes and add unit tests for task operations
