+++ api/controllers/tasks/create_test.go
@@ -1,0 +1,37 @@
+package tasks
+
+import (
+	"atomic_blend_api/models"
+	"bytes"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+)
+
+func TestCreateTask(t *testing.T) {
+	router, mockRepo := setupTest()
+
+	t.Run("successful create task", func(t *testing.T) {
+		task := createTestTask()
+
+		mockRepo.On("Create", mock.Anything, mock.AnythingOfType("*models.TaskEntity")).Return(task, nil)
+
+		taskJSON, _ := json.Marshal(task)
+		w := httptest.NewRecorder()
+		req, _ := http.NewRequest("POST", "/tasks", bytes.NewBuffer(taskJSON))
+		req.Header.Set("Content-Type", "application/json")
+		router.ServeHTTP(w, req)
+
+		assert.Equal(t, http.StatusCreated, w.Code)
+		var response models.TaskEntity
+		err := json.Unmarshal(w.Body.Bytes(), &response)
+		assert.NoError(t, err)
+		assert.Equal(t, task.Title, response.Title)
+		assert.Equal(t, task.StartDate, response.StartDate)
+		assert.Equal(t, task.EndDate, response.EndDate)
+	})
+}
