+++ api/controllers/tasks/get_by_id_test.go
@@ -1,0 +1,49 @@
+package tasks
+
+import (
+	"atomic_blend_api/models"
+	"encoding/json"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+)
+
+
+func TestGetTaskByID(t *testing.T) {
+	router, mockRepo := setupTest()
+
+	t.Run("successful get task by id", func(t *testing.T) {
+		taskID := primitive.NewObjectID().Hex()
+		task := createTestTask()
+		task.ID = taskID
+
+		mockRepo.On("GetByID", mock.Anything, taskID).Return(task, nil)
+
+		w := httptest.NewRecorder()
+		req, _ := http.NewRequest("GET", "/tasks/"+taskID, nil)
+		router.ServeHTTP(w, req)
+
+		assert.Equal(t, http.StatusOK, w.Code)
+		var response models.TaskEntity
+		err := json.Unmarshal(w.Body.Bytes(), &response)
+		assert.NoError(t, err)
+		assert.Equal(t, task.Title, response.Title)
+		assert.Equal(t, task.StartDate, response.StartDate)
+		assert.Equal(t, task.EndDate, response.EndDate)
+	})
+
+	t.Run("task not found", func(t *testing.T) {
+		taskID := primitive.NewObjectID().Hex()
+		mockRepo.On("GetByID", mock.Anything, taskID).Return(nil, nil)
+
+		w := httptest.NewRecorder()
+		req, _ := http.NewRequest("GET", "/tasks/"+taskID, nil)
+		router.ServeHTTP(w, req)
+
+		assert.Equal(t, http.StatusNotFound, w.Code)
+	})
+}
