+++ api/controllers/tasks/task_controller.go
@@ -4,6 +4,7 @@ import (
 	"atomic_blend_api/repositories"
 
 	"github.com/gin-gonic/gin"
+	"go.mongodb.org/mongo-driver/mongo"
 )
 
 // TaskController handles task related operations
@@ -19,13 +20,26 @@ func NewTaskController(taskRepo repositories.TaskRepositoryInterface) *TaskContr
 }
 
 // SetupRoutes sets up the task routes
-func (c *TaskController) SetupRoutes(router *gin.RouterGroup) {
+func SetupRoutes(router *gin.Engine, database *mongo.Database) {
+	taskRepo := repositories.NewTaskRepository(database)
+	taskController := NewTaskController(taskRepo)
+	setupTaskRoutes(router, taskController)
+}
+
+// SetupRoutesWithMock sets up the task routes with a mock repository for testing
+func SetupRoutesWithMock(router *gin.Engine, taskRepo repositories.TaskRepositoryInterface) {
+	taskController := NewTaskController(taskRepo)
+	setupTaskRoutes(router, taskController)
+}
+
+// setupTaskRoutes sets up the routes for task controller
+func setupTaskRoutes(router *gin.Engine, taskController *TaskController) {
 	taskRoutes := router.Group("/tasks")
 	{
-		taskRoutes.GET("", c.GetAllTasks)
-		taskRoutes.GET("/:id", c.GetTaskByID)
-		taskRoutes.POST("", c.CreateTask)
-		taskRoutes.PUT("/:id", c.UpdateTask)
-		taskRoutes.DELETE("/:id", c.DeleteTask)
+		taskRoutes.GET("", taskController.GetAllTasks)
+		taskRoutes.GET("/:id", taskController.GetTaskByID)
+		taskRoutes.POST("", taskController.CreateTask)
+		taskRoutes.PUT("/:id", taskController.UpdateTask)
+		taskRoutes.DELETE("/:id", taskController.DeleteTask)
 	}
 }
